Refactor CCKafkaClientsIntegrationTestHarness and dependent classes.
